refactor(settings): move search state into the states directory (@OmTiwariPandey)
